Move configuration-only data structures into `frontend::configuration` module
